requirements.txt
asgiref==3.5.2
Django==4.0.6
django-cors-headers==3.13.0
django-stubs==1.12.0
django-stubs-ext==0.5.0
djangorestframework==3.13.1
mypy==0.971
mypy-extensions==0.4.3
python-decouple==3.6
pytz==2022.1
sqlparse==0.4.2
tomli==2.0.1
types-pytz==2022.1.2
types-PyYAML==6.0.11
typing_extensions==4.3.0
tzdata==2022.1
